+++ example-usage/pages/subscriptions.tsx
@@ -1,4 +1,4 @@
-import { Box, Spinner, Stack } from "@chakra-ui/core";
+import { Box, Spinner, Stack } from "@chakra-ui/react";
 import { Hero, PageContainer, SectionTitle } from "landing-blocks";
 import React from "react";
 import { createClient } from "../hasura";
@@ -40,7 +40,7 @@ const Page = () => {
         )}
         {data && (
           <Stack spacing="20px">
-            {data?.user?.map((x) => (
+            {data?.user?.map((x: any) => (
               <Box borderRadius="10px" p="20px" borderWidth="1px">
                 {x.name}
               </Box>